Model and Sequential implemented

+++ dev/windows/aurora-cpp/tensor.cpp
@@ -27,7 +27,7 @@ void tensor::set(tensor _other) {
 tensor tensor::new_1d(size_t _a) {
 	tensor result;
 	result.resize(_a);
-	for (size_t i = 0; i < _a; i++) {
+	for (int i = 0; i < _a; i++) {
 		result[i] = 0;
 	}
 	return result;
@@ -36,7 +36,7 @@ tensor tensor::new_1d(size_t _a) {
 tensor tensor::new_2d(size_t _a, size_t _b) {
 	tensor result;
 	result.resize(_a);
-	for (size_t i = 0; i < _a; i++) 
+	for (int i = 0; i < _a; i++) 
 		result[i] = new_1d(_b);
 	
 	return result;
@@ -44,7 +44,7 @@ tensor tensor::new_2d(size_t _a, size_t _b) {
 
 tensor tensor::sum_1d() {
 	tensor result = 0;
-	for (size_t i = 0; i < size(); i++) 
+	for (int i = 0; i < size(); i++) 
 		result.val += at(i).val;
 	
 	return result;
@@ -79,7 +79,7 @@ tensor tensor::unroll() {
 	size_t w = width();
 	size_t h = height();
 	tensor result = new_1d(w * h);
-	for (size_t i = 0; i < h; i++)
+	for (int i = 0; i < h; i++)
 		for (size_t j = 0; j < w; j++)
 			result[i * w + j] = at(i)[j];
 	return result;
@@ -90,7 +90,7 @@ tensor tensor::roll(size_t _width) {
 	assert(s % _width == 0);
 	size_t h = s / _width;
 	tensor result = new_2d(h, _width);
-	for (size_t i = 0; i < size(); i++)
+	for (int i = 0; i < size(); i++)
 		result[i / _width][i % _width] = at(i);
 	return result;
 }
@@ -110,14 +110,14 @@ tensor tensor::row(size_t _a) {
 
 tensor tensor::col(size_t _a) {
 	tensor result = new_1d(size());
-	for (size_t i = 0; i < size(); i++)
+	for (int i = 0; i < size(); i++)
 		result[i] = at(i)[_a];
 	return result;
 }
 
 tensor tensor::range(size_t _start, size_t _len) {
 	tensor result = new_1d(_len);
-	for (size_t i = 0; i < _len; i++)
+	for (int i = 0; i < _len; i++)
 	{
 		size_t src = _start + i;
 		size_t dst = i;
@@ -133,14 +133,14 @@ tensor tensor::clone_row(size_t _a) {
 
 tensor tensor::clone_col(size_t _a) {
 	tensor result = new_1d(size());
-	for (size_t i = 0; i < size(); i++)
+	for (int i = 0; i < size(); i++)
 		result[i] = at(i)[_a];
 	return result.clone();
 }
 
 tensor tensor::clone_range(size_t _start, size_t _len) {
 	tensor result = new_1d(_len);
-	for (size_t i = 0; i < _len; i++)
+	for (int i = 0; i < _len; i++)
 	{
 		size_t src = _start + i;
 		size_t dst = i;
@@ -157,7 +157,7 @@ tensor tensor::add_1d(tensor _other) {
 
 void tensor::add_1d(tensor _other, tensor& _output) {
 	assert(size() == _other.size());
-	for (size_t i = 0; i < size(); i++)
+	for (int i = 0; i < size(); i++)
 		_output[i] = at(i).val + _other.at(i).val;
 }
 
@@ -169,7 +169,7 @@ tensor tensor::sub_1d(tensor _other) {
 
 void tensor::sub_1d(tensor _other, tensor& _output) {
 	assert(size() == _other.size());
-	for (size_t i = 0; i < size(); i++)
+	for (int i = 0; i < size(); i++)
 		_output[i] = at(i).val - _other.at(i).val;
 }
 
@@ -181,7 +181,7 @@ tensor tensor::mul_1d(tensor _other) {
 
 void tensor::mul_1d(tensor _other, tensor& _output) {
 	assert(size() == _other.size());
-	for (size_t i = 0; i < size(); i++)
+	for (int i = 0; i < size(); i++)
 		_output[i] = at(i).val * _other.at(i).val;
 }
 
@@ -193,7 +193,7 @@ tensor tensor::div_1d(tensor _other) {
 
 void tensor::div_1d(tensor _other, tensor& _output) {
 	assert(size() == _other.size());
-	for (size_t i = 0; i < size(); i++)
+	for (int i = 0; i < size(); i++)
 		_output[i] = at(i).val / _other.at(i).val;
 }
 
@@ -206,7 +206,7 @@ tensor tensor::dot_1d(tensor _other) {
 void tensor::dot_1d(tensor _other, tensor& _output) {
 	assert(size() == _other.size());
 	_output.val = (double)0;
-	for (size_t i = 0; i < size(); i++)
+	for (int i = 0; i < size(); i++)
 		_output.val += at(i).val * _other.at(i).val;
 }
 
@@ -218,7 +218,7 @@ tensor tensor::add_2d(tensor _other) {
 
 void tensor::add_2d(tensor _other, tensor& _output) {
 	assert(size() == _other.size());
-	for (size_t i = 0; i < size(); i++)
+	for (int i = 0; i < size(); i++)
 		_output[i] = at(i).add_1d(_other[i]);
 }
 
@@ -230,7 +230,7 @@ tensor tensor::sub_2d(tensor _other) {
 
 void tensor::sub_2d(tensor _other, tensor& _output) {
 	assert(size() == _other.size());
-	for (size_t i = 0; i < size(); i++)
+	for (int i = 0; i < size(); i++)
 		_output[i] = at(i).sub_1d(_other[i]);
 }
 
@@ -242,7 +242,7 @@ tensor tensor::mul_2d(tensor _other) {
 
 void tensor::mul_2d(tensor _other, tensor& _output) {
 	assert(size() == _other.size());
-	for (size_t i = 0; i < size(); i++)
+	for (int i = 0; i < size(); i++)
 		_output[i] = at(i).mul_1d(_other[i]);
 }
 
@@ -254,7 +254,7 @@ tensor tensor::div_2d(tensor _other) {
 
 void tensor::div_2d(tensor _other, tensor& _output) {
 	assert(size() == _other.size());
-	for (size_t i = 0; i < size(); i++)
+	for (int i = 0; i < size(); i++)
 		_output[i] = at(i).div_1d(_other[i]);
 }
 
@@ -266,7 +266,7 @@ tensor tensor::dot_2d(tensor _other) {
 
 void tensor::dot_2d(tensor _other, tensor& _output) {
 	assert(width() == _other.height());
-	for (size_t i = 0; i < height(); i++)
+	for (int i = 0; i < height(); i++)
 		for (size_t j = 0; j < _other.width(); j++)
 			_output[i][j] = at(i).dot_1d(_other.col(j));
 }
@@ -274,27 +274,27 @@ void tensor::dot_2d(tensor _other, tensor& _output) {
 void tensor::clone(tensor& _output) {
 	assert(size() == _output.size());
 	_output.val = val;
-	for (size_t i = 0; i < size(); i++)
+	for (int i = 0; i < size(); i++)
 		at(i).clone(_output.at(i));
 }
 
 void tensor::link(tensor& _other) {
 	val_ptr.link(_other.val_ptr);
 	val = val_ptr.get();
-	for (size_t i = 0; i < size(); i++)
+	for (int i = 0; i < size(); i++)
 		at(i).link(_other.at(i));
 }
 
 void tensor::unlink() {
 	val_ptr.unlink();
-	for (size_t i = 0; i < size(); i++)
+	for (int i = 0; i < size(); i++)
 		at(i).unlink();
 }
 
 tensor tensor::clone() {
 	tensor result = tensor(val);
 	result.resize(size());
-	for (size_t i = 0; i < size(); i++)
+	for (int i = 0; i < size(); i++)
 		result.at(i) = at(i).clone();
 	return result;
 }
@@ -303,7 +303,7 @@ string tensor::to_string() {
 	if (size() == 0)
 		return std::to_string(val);
 	string result = "[";
-	for (size_t i = 0; i < size() - 1; i++)
+	for (int i = 0; i < size() - 1; i++)
 		result += at(i).to_string() + " ";
 	result += at(size() - 1).to_string() + "]";
 	return result;
@@ -319,7 +319,7 @@ void tensor::operator=(tensor _other) {
 
 void tensor::clear() {
 	val = (double)0;
-	for (size_t i = 0; i < size(); i++)
+	for (int i = 0; i < size(); i++)
 		at(i).clear();
 }
